pyproject.toml
[project]
name = "ocrformduel"
version = "0.1.0"
description = "W2 Vision Quest"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "opencv-python>=4.13.0.90",
    "paddleocr>=3.4.0",
    "pillow>=12.1.0",
    "pytesseract>=0.3.13",
]
